+++ src/anemoi/inference/output.py
@@ -6,26 +6,67 @@
 # granted to it by virtue of its status as an intergovernmental organisation
 # nor does it submit to any jurisdiction.
 #
+import logging
 from abc import ABC
 from abc import abstractmethod
+
+LOG = logging.getLogger(__name__)
 
 
 class Output(ABC):
     """_summary_"""
 
-    def __init__(self, context):
+    def __init__(self, context, output_frequency=None, write_initial_step=False):
+        from anemoi.utils.dates import as_timedelta
+
         self.context = context
         self.checkpoint = context.checkpoint
+        self.reference_date = None
+
+        self.write_step_zero = write_initial_step and context.write_initial_step
+
+        self.output_frequency = output_frequency or context.output_frequency
+        if self.output_frequency is not None:
+            self.output_frequency = as_timedelta(self.output_frequency)
 
     def __repr__(self):
         return f"{self.__class__.__name__}()"
 
-    @abstractmethod
     def write_initial_state(self, state):
-        pass
+        self._init(state)
+        if self.write_step_zero:
+            return self.write_initial_step(state, state["date"] - self.reference_date)
+
+    def write_state(self, state):
+        self._init(state)
+
+        step = state["date"] - self.reference_date
+        if self.output_frequency is not None:
+            if (step % self.output_frequency).total_seconds() != 0:
+                return
+
+        return self.write_step(state, step)
+
+    def _init(self, state):
+        if self.reference_date is not None:
+            return
+
+        self.reference_date = state["date"]
+
+        self.open(state)
+
+    def write_initial_step(self, state, step):
+        """This method should not be called directly
+        call `write_initial_state` instead.
+        """
+        reduced_state = self.reduce(state)
+        self.write_step(reduced_state, step)
 
     @abstractmethod
-    def write_state(self, state):
+    def write_step(self, state, step):
+        """This method should be be called directly
+        call `write_state` instead.
+        """
         pass
 
     def reduce(self, state):
@@ -35,6 +76,10 @@ class Output(ABC):
         for field, values in state["fields"].items():
             reduced_state["fields"][field] = values[-1, :]
         return reduced_state
+
+    def open(self, state):
+        # Override this method when initialisation is needed
+        pass
 
     def close(self):
         pass