+++ src/anemoi/inference/outputs/extract_lam.py
@@ -22,8 +22,9 @@ LOG = logging.getLogger(__name__)
 class ExtractLamOutput(Output):
     """_summary_"""
 
-    def __init__(self, context, *, output, points="cutout_mask"):
-        super().__init__(context)
+    def __init__(self, context, *, output, points="cutout_mask", output_frequency=None, write_initial_step=False):
+        super().__init__(context, output_frequency=output_frequency, write_initial_step=write_initial_step)
+
         if isinstance(points, str):
             mask = self.checkpoint.load_supporting_array(points)
             points = -np.sum(mask)  # This is the global, we want the lam
@@ -34,11 +35,11 @@ class ExtractLamOutput(Output):
     def __repr__(self):
         return f"ExtractLamOutput({self.points}, {self.output})"
 
-    def write_initial_state(self, state):
-        self.output.write_initial_state(self._apply_mask(state))
+    def write_initial_step(self, state, step):
+        self.output.write_initial_step(self._apply_mask(state), step)
 
-    def write_state(self, state):
-        self.output.write_state(self._apply_mask(state))
+    def write_step(self, state, step):
+        self.output.write_step(self._apply_mask(state), step)
 
     def _apply_mask(self, state):